Integrate canonical gitignore template from github
The canonical gitignore template can be found at
https://github.com/github/gitignore/blob/master/Python.gitignore
and it includes rules to ignore many files that might be generated
for a python project but should not be checked into version control.

The rules from the previous gitignore are also included. Those rules
are also simplified. For example **/*.ext is equivalent to *.ext.